+++ resources/[dg]/dg-vehicles/typescript/src/client/modules/stances/service.stances.ts
@@ -4,7 +4,7 @@ import { NOTIFICATION_ID } from './constants.stances';
 const closeVehicles: Map<number, Stance.Data> = new Map();
 let stancingThread: NodeJS.Timer | null = null;
 
-export const handleStanceStateUpdate = (netId: number, stanceData: Stance.Data) => {
+export const handleStanceStateUpdate = (netId: number, stanceData: Stance.Data | null) => {
   // handler fires before entity exists for client. This handler is used for current vehicle only so we dont need to await
   if (!NetworkDoesNetworkIdExist(netId) || !NetworkDoesEntityExistWithNetworkId(netId)) return false;
   const veh = NetworkGetEntityFromNetworkId(netId);
@@ -14,7 +14,12 @@ export const handleStanceStateUpdate = (netId: number, stanceData: Stance.Data) 
   return true;
 };
 
-export const setCloseVehicleStance = (vehicle: number, stanceData: Stance.Data) => {
+export const setCloseVehicleStance = (vehicle: number, stanceData: Stance.Data | null) => {
+  if (!stanceData) {
+    closeVehicles.delete(vehicle);
+    return;
+  }
+
   closeVehicles.set(vehicle, stanceData);
 
   if (closeVehicles.size !== 0) {